Remove unnecessary {block,..} expressions in do([]) transformation
The previous commit introduced {block,.. } expressions (begin/end blocks) to fix the problem with 'match' expression being the first in 'do'. But as result every occurrence of 'match' in 'do' would result in {block, ..} even if it was unnecessary.
With this fix {block,..} is now added only when it is necessary.